+++ src/components/welcomePage/github-Icon/index.tsx
@@ -1,0 +1,1 @@
+export { GitHubIcon } from './ghicon';
